fix: 积分改为精确小数累加并清理诊断步骤
- UserEntUsage 逐包取整改为保留小数累加，末尾保留两位小数，与界面总可用积分一致(如4855.36)

- 积分列输出 %.2f

- signin.yml 移除临时诊断步骤与 push 触发，恢复生产配置

+++ cmd/signin/main.go
@@ -20,7 +20,7 @@ type row struct {
 	nick   string
 	status string
 	detail string
-	remain int64
+	remain float64
 	hasRem bool
 }
 
@@ -121,7 +121,7 @@ func main() {
 	for _, r := range rows {
 		remain := "-"
 		if r.hasRem {
-			remain = fmt.Sprintf("%d", r.remain)
+			remain = fmt.Sprintf("%.2f", r.remain)
 		}
 		fmt.Printf("│ %-36s │ %-13s │ %-12s │ %-8s │ %-36s │\n",
 			trunc(r.uid, 36), trunc(r.nick, 13), r.status, remain, trunc(r.detail, 36))
